refactor(reposicao): extrai AdminReposicaoPedidos god-component (1571→236)
Quebra o god-component AdminReposicaoPedidos.tsx (1571 linhas) em
subcomponentes sob src/components/reposicao/pedidos/. Extração pura,
behavior-preserving (zero mudança de lógica/JSX/queries):

- types.ts        — Status, StatusEnvioPortal, PedidoSugerido, CondicaoPagamento, PedidoItem
- shared.ts       — EMPRESA, statusMeta, portalStatusMeta, getEstoqueZoneClass, formatBRL, formatTime
- badges.tsx      — StatusBadge, SplitInfo, PortalBadge
- CycleIndicator.tsx, PortalDrawer.tsx, PedidoRow.tsx, CancelarModal.tsx, CiclosAnteriores.tsx
- DetalhesModal.tsx — DetalhesModal + PortalStatusPanel + HistoricoAcoesPanel (co-locados; só usados aqui)

A página vira orquestrador (236 linhas): estado + queries + layout, importando os subcomponentes.
Split shared.ts (não-componentes) vs badges.tsx (componentes) resolve o aviso react-refresh.

Primeiro de uma série de refactors de god-components (§10). Revisado via codex review
(extração fiel, sem bug introduzido). typecheck:strict + tsc + 392/392 testes + lint 0.
