+++ src/components/reposicao/pedidos/PedidoRow.tsx
@@ -1,0 +1,95 @@
+import { Button } from '@/components/ui/button';
+import { TableCell, TableRow } from '@/components/ui/table';
+import { Eye, ExternalLink, Loader2, XCircle, Zap } from 'lucide-react';
+import { format } from 'date-fns';
+import { PedidoSugerido } from './types';
+import { formatBRL, formatTime } from './shared';
+import { StatusBadge, SplitInfo, PortalBadge } from './badges';
+
+export function PedidoRow({
+  p,
+  onVerDetalhes,
+  onCancelar,
+  onVerPortal,
+  onDisparar,
+  disparando,
+}: {
+  p: PedidoSugerido;
+  onVerDetalhes: () => void;
+  onCancelar: () => void;
+  onVerPortal: () => void;
+  onDisparar: () => void;
+  disparando: boolean;
+}) {
+  const podeAprovar = p.status === 'pendente_aprovacao' || p.status === 'bloqueado_guardrail';
+  const podeCancelar = ['pendente_aprovacao', 'bloqueado_guardrail', 'aprovado_aguardando_disparo'].includes(p.status);
+  const podeDisparar = p.status === 'aprovado_aguardando_disparo';
+
+  const showAprovacao = p.status === 'aprovado_aguardando_disparo' || p.status === 'disparado';
+
+  return (
+    <TableRow className={p.status === 'bloqueado_guardrail' ? 'bg-destructive/5' : ''}>
+      <TableCell>
+        <div className="flex flex-wrap items-center gap-1">
+          <StatusBadge status={p.status} />
+          <SplitInfo pedido={p} />
+        </div>
+      </TableCell>
+      <TableCell>
+        <div className="font-medium">{p.fornecedor_nome}</div>
+        <div className="text-xs text-muted-foreground">{p.grupo_codigo ?? '—'}</div>
+      </TableCell>
+      <TableCell className="text-right tabular-nums">{p.num_skus}</TableCell>
+      <TableCell className="text-right tabular-nums font-medium">{formatBRL(p.valor_total)}</TableCell>
+      <TableCell className="text-right tabular-nums">
+        {p.delta_vs_anterior_perc !== null ? (
+          <span className={Number(p.delta_vs_anterior_perc) >= 0 ? 'text-status-success' : 'text-destructive'}>
+            {Number(p.delta_vs_anterior_perc) >= 0 ? '+' : ''}{Number(p.delta_vs_anterior_perc).toFixed(1)}%
+          </span>
+        ) : <span className="text-muted-foreground">—</span>}
+      </TableCell>
+      <TableCell className="text-right">{formatTime(p.horario_corte_planejado)}</TableCell>
+      <TableCell>
+        <PortalBadge pedido={p} onClick={onVerPortal} />
+      </TableCell>
+      <TableCell className="text-xs">
+        {showAprovacao && p.aprovado_em ? (
+          <div>
+            <div className="font-medium tabular-nums">{format(new Date(p.aprovado_em), 'dd/MM HH:mm')}</div>
+            <div className="text-muted-foreground line-clamp-1">{p.aprovado_por ?? '—'}</div>
+          </div>
+        ) : (
+          <span className="text-muted-foreground">—</span>
+        )}
+      </TableCell>
+      <TableCell>
+        <div className="flex justify-end gap-1">
+          <Button size="sm" variant="ghost" onClick={onVerDetalhes}>
+            <Eye className="w-4 h-4 mr-1" />Detalhes
+          </Button>
+          {podeAprovar && (
+            <Button size="sm" variant="default" onClick={onVerDetalhes}>Aprovar</Button>
+          )}
+          {podeDisparar && (
+            <Button size="sm" variant="default" onClick={onDisparar} disabled={disparando}>
+              {disparando ? <Loader2 className="w-4 h-4 mr-1 animate-spin" /> : <Zap className="w-4 h-4 mr-1" />}
+              Disparar
+            </Button>
+          )}
+          {podeCancelar && (
+            <Button size="sm" variant="outline" onClick={onCancelar}>
+              <XCircle className="w-4 h-4" />
+            </Button>
+          )}
+          {p.status === 'disparado' && p.omie_pedido_compra_numero && (
+            <Button size="sm" variant="ghost" asChild>
+              <a href={`https://app.omie.com.br/`} target="_blank" rel="noreferrer">
+                <ExternalLink className="w-4 h-4 mr-1" />Omie
+              </a>
+            </Button>
+          )}
+        </div>
+      </TableCell>
+    </TableRow>
+  );
+}
